requirements.txt
python-telegram-bot==22.7
openpyxl
python-pptx
Pillow
fpdf2
